+++ core-ng/src/main/java/core/framework/impl/template/node/Attributes.java
@@ -19,7 +19,7 @@ public class Attributes {
         attributes.put(attribute.name, attribute);
     }
 
-    public void buildTemplate(ContainerFragment parent, TemplateMetaContext context) {
+    void buildTemplate(ContainerFragment parent, TemplateMetaContext context) {
         validate();
 
         for (Attribute attribute : attributes.values()) {
@@ -53,14 +53,14 @@ public class Attributes {
         return !attribute.isDynamic() && (attributes.containsKey("c:" + name) || attributes.containsKey("m:" + name));   // there is dynamic attribute to overwrite
     }
 
-    public boolean containDynamicContent() {
+    boolean containDynamicContent() {
         return attributes.containsKey("c:text")
             || attributes.containsKey("m:text")
             || attributes.containsKey("c:html")
             || attributes.containsKey("c:include");
     }
 
-    public Attribute dynamicContentAttribute() {
+    Attribute dynamicContentAttribute() {
         Attribute attribute = attributes.get("c:text");
         if (attribute != null) return attribute;
         attribute = attributes.get("m:text");
@@ -70,7 +70,7 @@ public class Attributes {
         return attributes.get("c:include");
     }
 
-    public List<Attribute> flowAttributes() {
+    List<Attribute> flowAttributes() {
         List<Attribute> attributes = new ArrayList<>();
         this.attributes.forEach((name, attr) -> {
             if ("c:if".equals(name) || "c:for".equals(name)) attributes.add(attr);