+++ core-ng/src/main/java/core/framework/web/site/Message.java
@@ -1,14 +1,12 @@
 package core.framework.web.site;
 
-import java.util.Optional;
-
 /**
  * @author neo
  */
 public interface Message {
-    Optional<String> get(String key, String language);
+    String get(String key, String language);
 
-    default Optional<String> get(String key) {
+    default String get(String key) {
         return get(key, null);
     }
 }