+++ core-ng/src/test/java/core/framework/impl/web/site/MessageImplTest.java
@@ -1,11 +1,13 @@
 package core.framework.impl.web.site;
 
+import core.framework.util.Properties;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * @author neo
@@ -22,24 +24,54 @@ class MessageImplTest {
 
     @Test
     void language() {
-        assertEquals("en_US", message.language("messages_en_US.properties"));
-        assertEquals("en", message.language("messages_en.properties"));
+        assertThat(message.language("messages_en_US.properties")).isEqualTo("en_US");
+        assertThat(message.language("messages_en.properties")).isEqualTo("en");
 
-        assertEquals(MessageImpl.DEFAULT_LANGUAGE, message.language("messages.properties"));
+        assertThat(message.language("messages.properties")).isEqualTo(MessageImpl.DEFAULT_LANGUAGE);
+
+        assertThatThrownBy(() -> message.language("invalid.message_E.properties"))
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("property path must match 'path/name_language.properties' pattern");
     }
 
     @Test
-    void message() {
-        assertEquals("value1", message.get("key1", "zh").orElseThrow());
-        assertEquals("value1", message.get("key1", "en").orElseThrow());
-        assertEquals("value1", message.get("key1", "en_US").orElseThrow());
+    void get() {
+        assertThat(message.get("key1", "zh")).isEqualTo("value1");
+        assertThat(message.get("key1", "en")).isEqualTo("value1");
+        assertThat(message.get("key1", "en_US")).isEqualTo("value1");
 
-        assertEquals("value2", message.get("key2", "zh").orElseThrow());
-        assertEquals("en_value2", message.get("key2", "en").orElseThrow());
-        assertEquals("en_value2", message.get("key2", "en_US").orElseThrow());
+        assertThat(message.get("key2", "zh")).isEqualTo("value2");
+        assertThat(message.get("key2", "en")).isEqualTo("en_value2");
+        assertThat(message.get("key2", "en_US")).isEqualTo("en_value2");
 
-        assertEquals("value3", message.get("key3", "zh").orElseThrow());
-        assertEquals("en_value3", message.get("key3", "en").orElseThrow());
-        assertEquals("en_US_value3", message.get("key3", "en_US").orElseThrow());
+        assertThat(message.get("key3", "zh")).isEqualTo("value3");
+        assertThat(message.get("key3", "en")).isEqualTo("en_value3");
+        assertThat(message.get("key3", "en_US")).isEqualTo("en_US_value3");
+    }
+
+    @Test
+    void getWithNotDefinedLanguage() {
+        assertThatThrownBy(() -> message.get("key1", "es_GT"))
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("language is not defined");
+    }
+
+    @Test
+    void getWithNotExistedKey() {
+        assertThatThrownBy(() -> message.get("notExistedKey", "zh"))
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("can not find message");
+    }
+
+    @Test
+    void validateMessageKeys() {
+        Properties properties = new Properties();
+        properties.set("key4", "value4");
+        message.messages.get("zh").add(properties);
+
+        assertThatThrownBy(() -> message.validateMessageKeys())
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("message keys are missing for language")
+                .hasMessageContaining("language=en_US");
     }
 }